+++ CHANGELOG.md
@@ -5,10 +5,10 @@
 [Full Changelog](https://github.com/SeldonIO/alibi-detect/compare/v0.10.4...v0.10.5)
 
 ### Fixed
-- Fixed a bug preventing backward compatibility when loading detectors (containing TensorFlow models) saved with `alibi-detect<0.10.0`
+- Fixed a bug preventing backward compatibility when loading detectors (containing TensorFlow models) saved with `<v0.10.0`
 ([#729](https://github.com/SeldonIO/alibi-detect/pull/729)). This bug also meant that detectors (containing TensorFlow models) saved with 
-`save_detector(..., legacy=True)` in `alibi-detect>=0.10.0` did not properly obey the legacy file format. The `config.toml` file format used by 
-default in `alibi-detect>=0.10.0` is not affected. 
+`save_detector(..., legacy=True)` in `>=v0.10.0` did not properly obey the legacy file format. The `config.toml` file format used by 
+default in `>=v0.10.0` is not affected. 
 
 ## v0.10.4
 ## [v0.10.4](https://github.com/SeldonIO/alibi-detect/tree/v0.10.4) (2022-10-21)