+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## v0.10.5
+## [v0.10.5](https://github.com/SeldonIO/alibi-detect/tree/v0.10.5) (2023-01-24)
+[Full Changelog](https://github.com/SeldonIO/alibi-detect/compare/v0.10.4...v0.10.5)
+
+### Fixed
+- Fixed a bug preventing backward compatibility when loading detectors (containing TensorFlow models) saved with `<v0.10.0`
+([#729](https://github.com/SeldonIO/alibi-detect/pull/729)). This bug also meant that detectors (containing TensorFlow models) saved with 
+`save_detector(..., legacy=True)` in `>=v0.10.0` did not properly obey the legacy file format. The `config.toml` file format used by 
+default in `>=v0.10.0` is not affected. 
+
 ## v0.10.4
 ## [v0.10.4](https://github.com/SeldonIO/alibi-detect/tree/v0.10.4) (2022-10-21)
 [Full Changelog](https://github.com/SeldonIO/alibi-detect/compare/v0.10.3...v0.10.4)